Extraer cursos recomendados Extraida seccion cursos recomendados
